+++ tests/GovUk.Frontend.AspNetCore.Tests/ComponentGeneration/ComponentFixtureData.cs
@@ -25,6 +25,7 @@ public class ComponentFixtureData(
         _serializerOptions.Converters.Add(new AttributeCollectionJsonConverter());
         _serializerOptions.Converters.Add(new TemplateStringJsonConverter());
         _serializerOptions.Converters.Add(new TableOptionsRowJsonConverter());
+        _serializerOptions.Converters.Add(new ServiceNavigationOptionsEndSlotJsonConverter());
 
     }